+++ src/Api/Dfe.Complete.Api/Controllers/ProjectGroupController.cs
@@ -5,6 +5,7 @@ using Swashbuckle.AspNetCore.Annotations;
 using Dfe.Complete.Application.Projects.Models;
 using Microsoft.AspNetCore.Authorization;
 using Dfe.Complete.Application.Projects.Queries.GetProject;
+using Dfe.Complete.Application.ProjectGroups.Commands;
 
 namespace Dfe.Complete.Api.Controllers
 {
@@ -14,6 +15,24 @@ namespace Dfe.Complete.Api.Controllers
     [Route("v{version:apiVersion}/[controller]")]
     public class ProjectGroupController(ISender sender) : ControllerBase
     {
+        /// <summary>
+        /// Creates a project group.
+        /// </summary>
+        /// <param name="request">The request.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWrite")]
+        [HttpPost]
+        [SwaggerResponse(201, "Project group created successfully.", typeof(Guid))]
+        [SwaggerResponse(400, "Invalid request data.")]
+        public async Task<IActionResult> CreateProjectGroupAsync([FromBody] CreateProjectGroupCommand request, CancellationToken cancellationToken)
+        {
+            var result = await sender.Send(request, cancellationToken);
+
+            if (!result.IsSuccess || result.Value == null)
+                return BadRequest(result.Error);
+            return Created("", result.Value!.Value);
+        }
+
         /// <summary>
         /// Gets the Project group by Id.
         /// </summary>
@@ -29,7 +48,7 @@ namespace Dfe.Complete.Api.Controllers
 
             return Ok(projectGroup.Value);
         }
-        
+
         /// <summary>
         /// Gets a list of the projects group and included establishments.
         /// </summary>