+++ CHANGELOG.md
@@ -86,6 +86,8 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - Added 'Land consent letter task for transfer projects'
 - Adding key contact record on confirming project handover and logging error message if key contact is already been added.
 - Added `Check accuracy of high needs places information` task for conversion project.
+- Added 'Land consent letter task for transfer projects'
+- Added project group creation end point
 
 ### Fixed